// src/types.ts
export interface Patient {
  id?: string
  room: string
  name: string
  mrn: string
  dob: string
  age: string
  admit_date: string
  ht: string
  wt: string
  dx: string
  md: string
  consults: string
  parent_guardian: string
  hx: string
  diet: string
  activity: string
  ivf: string
  ivf_site: string
  meds: string
  prn: string
  allergies: string
  neb_tx: string
  code_status: string
  precautions: string
  notes: string
  labs: string
  echo: string
  cxr: string
  io: string
  special_equipment: string
  new_meds: string
  new_concerns: string
  pending_labs: string
  pending_procedures: string
  pending_consults: string
  plan_of_care: string
  other: string
  pain_reassessment: string
  restraints: string
  suicide_level: string
  critical_results: string
  pews: string
  hrfe: string
  home_meds: string
  cx_gi_prop: string
  cx_sq_heparin: string
  cx_enoxaparin: string
  cx_dvt_boots: string
  cx_aspirin: string
  cx_bowel_reg: string
  cx_oral_care: string
  cx_hob: string
  cx_foley_dc: string
  cx_lines_dc: string
  prophylaxis_notes: string
  cx_pt: string
  cx_ot: string
  cx_speech: string
  cx_diet: string
  cx_sw: string
  cx_cm: string
  cx_chaplain: string
  cx_wound: string
  cx_palliative: string
  cx_pharmacy: string
  consult_notes: string
  dc_target_dispo: string
  cx_dme: string
  cx_ssoc: string
  cx_home_health: string
  cx_snf: string
  cx_fu_md: string
  cx_pt_edu: string
  cx_ins_auth: string
  cx_transport: string
  cx_rx_dc: string
  cx_dc_summary: string
  dc_notes: string
  scriba_patient_id: string
  solo_encounter_id: string
  created_at?: string
  updated_at?: string
}

export const EMPTY_PATIENT: Patient = {
  room: '',
  name: '',
  mrn: '',
  dob: '',
  age: '',
  admit_date: '',
  ht: '',
  wt: '',
  dx: '',
  md: '',
  consults: '',
  parent_guardian: '',
  hx: '',
  diet: '',
  activity: '',
  ivf: '',
  ivf_site: '',
  meds: '',
  prn: '',
  allergies: '',
  neb_tx: '',
  code_status: '',
  precautions: '',
  notes: '',
  labs: '',
  echo: '',
  cxr: '',
  io: '',
  special_equipment: '',
  new_meds: '',
  new_concerns: '',
  pending_labs: '',
  pending_procedures: '',
  pending_consults: '',
  plan_of_care: '',
  other: '',
  pain_reassessment: '',
  restraints: '',
  suicide_level: '',
  critical_results: '',
  pews: '',
  hrfe: '',
  home_meds: '',
  cx_gi_prop: '',
  cx_sq_heparin: '',
  cx_enoxaparin: '',
  cx_dvt_boots: '',
  cx_aspirin: '',
  cx_bowel_reg: '',
  cx_oral_care: '',
  cx_hob: '',
  cx_foley_dc: '',
  cx_lines_dc: '',
  prophylaxis_notes: '',
  cx_pt: '',
  cx_ot: '',
  cx_speech: '',
  cx_diet: '',
  cx_sw: '',
  cx_cm: '',
  cx_chaplain: '',
  cx_wound: '',
  cx_palliative: '',
  cx_pharmacy: '',
  consult_notes: '',
  dc_target_dispo: '',
  cx_dme: '',
  cx_ssoc: '',
  cx_home_health: '',
  cx_snf: '',
  cx_fu_md: '',
  cx_pt_edu: '',
  cx_ins_auth: '',
  cx_transport: '',
  cx_rx_dc: '',
  cx_dc_summary: '',
  dc_notes: '',
  scriba_patient_id: '',
  solo_encounter_id: '',
}

export interface ParsedCensusData {
  patients: Partial<Patient>[]
  error?: string
}